backend/requirements.txt
fastapi[standard]
Pillow
pdf2docx
python-magic